Adjustments - Documents and News

+++ database/migrations/2022_02_09_150321_create_projetos_table.php
@@ -21,7 +21,7 @@ class CreateProjetosTable extends Migration
             $table->string('descricao');
             $table->string('link');
             $table->string('img');
-            $table->integer('userID')->unsigned();
+            $table->bigInteger('userID')->unsigned();
             $table->foreign('userID')->references('id')->on('users');
             $table->timestamps();
         });
